+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -241,5 +241,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-27T09:34:54.320Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-27T11:28:47.196Z")
     r"""The date and time the resource was last updated"""