+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8JGAVXE2SYCR7P0JPW1PVDE"
+        "01K8JPVD5CWWZ3XST5YS56WGH7"
     )
     r"""The unique identifier of the data source"""