+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-27T09:34:57.398Z")
+    created: Optional[datetime] = parse_datetime("2025-10-27T11:28:50.372Z")
     r"""The date and time the resource was created"""