+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -161,5 +161,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-27T09:34:54.320Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-27T11:28:47.196Z")
     r"""The date and time the resource was last updated"""