+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8JGAVXC8WP9RV112H6E6ZEC"
+        "01K8JPVD5B2NBJA6QBY0MYV6FV"
     )
     r"""The unique identifier of the data source"""