+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8JGAVXR2EXM4WYRK3PTVDDA"
+        "01K8JPVD5E1Y6M932QAAAZG0DK"
     )
     r"""The unique identifier of the data source"""